+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.1] - 2026-06-04
+
+- [@JShull](https://github.com/jshull).
+  - License Updates
+
 ## [0.7.0] - 2026-04-29
 
 ### 0.7.0 Added